Update body.tex
In the second paragraph, I distinguish the 'annual planning process'  (meaning in part the POP process) from the twice a year detail planning we have ADs and TLs do to develop L3 and Epics.  This detail is provided later in the document and should be an overview here.  Does "Leadership Team" mean ADs and TLs?