Update body.tex
In section 2.3 Work Breakdown Structure, updates to the WBS table include indenting and adding some new elements.  This numbering corresponds to the AURA/CAS Costpoint accounting project codes, but there is now a facility to map those numbers into an alternate numbering system which we could set up to match to the Rubin suite of budgeting tools.  Do we want to update the numbering here to match, e.g. the WBS Labor tab? Also, more changes will be needed to update the restructure of RDP and possibly ROO.