Update body.tex
Need to introduce and build out the Team Effort Plan as a way to estimate each Team's total time to allocate.  Also, this still refers to Construction conventions for hours estimation.  We will need to confirm with Wil and Leanne that we will use the NOIRLab numbers and the Team Effort Plan to do this in Operations.