Update body.tex
Defining the Schedule with Milestones:  we will need to define a convention for Start Date.  Since DP and SF do not historically use this field, we have typically been setting the Start Date in Smartsheet as the first day of the FY.  We need to confirm RO Milestone ID is not used elsewhere before removing it from the Milestone issue type.